Move executeCommand to common
